Added Java implementation of the benchmark suite (#49)
Implemented the main benchmarks present in the Python suite in Java. The class Benchmark.java provides a template for future use cases. Created configuration files to easily adapt the benchmarks to different types of workloads.
The suite is built as a Maven project, indeed running mvn test will execute all the benchmarks based on the configuration file of the appropriate GPU architecture.